Initial Open Challenge Commit
NoxOpenChallengeBehavior is based on the Behavior module. Motions LayDown
and StandUp are added. The NoxOpenchallenge behavior, goes straight to the
ball (it does not turn the robot if the ball was not detected),
fine approach it, and then execute the special action routine.
NoxOpenchallenge is polling the  openChallenge config to update
its parameters on the fly.